cleanup

+++ streamer.py
@@ -37,26 +37,28 @@ def main():
         logger.info("latest committed offset token before insert: " + (latest_committed_offset_token or "None"))
         if not latest_committed_offset_token:
             latest_committed_offset_token = 0
-        # send data with batching of 20 rows per batch
+        # send data with batching
         batch_size = 20
         logger.info("start sending insert rows with batching")
-        con = sqlite3.connect("/mnt/data/data.db")
+        con = sqlite3.connect("data.db")
         while True:
             cur = con.cursor()
             rows = []
+            last_offset_token = None
             for row in cur.execute("SELECT * FROM tdata WHERE id > ? ORDER BY id ASC LIMIT ?", (int(latest_committed_offset_token), batch_size)):
                 rows.append(row[1])
-                offset_token = row[0]
-            nl_json = "\n".join(rows)
-            channel.insert_rows(nl_json, offset_token=str(offset_token))
-            logger.info("inserted rows in batch to offset token: " + str(offset_token))
-            # cleanup from committed offset
-            current_committed_offset_token = channel.get_latest_committed_offset_token()
-            if current_committed_offset_token and int(current_committed_offset_token) > int(latest_committed_offset_token):
-                logger.info("delete from latest committed offset token: " + current_committed_offset_token)
-                cur.execute("DELETE FROM tdata WHERE id <= ?", (int(current_committed_offset_token),))
-                con.commit()
-                latest_committed_offset_token = current_committed_offset_token
+                last_offset_token = row[0]
+            if len(rows) > 0 and last_offset_token:
+                nl_json = "\n".join(rows)
+                channel.insert_rows(nl_json, offset_token=str(last_offset_token))
+                logger.info("inserted rows in batch to offset token: " + str(last_offset_token))
+                # cleanup from committed offset
+                current_committed_offset_token = channel.get_latest_committed_offset_token()
+                if current_committed_offset_token and int(current_committed_offset_token) > int(latest_committed_offset_token):
+                    logger.info("delete from latest committed offset token: " + current_committed_offset_token)
+                    cur.execute("DELETE FROM tdata WHERE id <= ?", (int(current_committed_offset_token),))
+                    con.commit()
+                    latest_committed_offset_token = current_committed_offset_token
      
     
 if __name__ == "__main__":
